CPU profiler with code hotspots support (#23)
This adds a separate CPU profiler which supports sample labels for code
hotspots and endpoint aggregation support. It works on wall profile timing,
however wall profile data is not included as the intent is to keep it separate
for now as it loses inlining data so the built-in wall profile is still better for
wall profiling data. Hopefully we can eventually get an API added to V8 to
get inlining information.