Add OOM Monitoring option to heap profiler (#82)
v8 allows to register a callback with AddNearHeapLimitCallback that is invoked when v8 is about to OOM.
In this callback we are in a GC context, and we cannot invoke any JS synchronously.
We can extend the heap size by a given amount (the return value of the callback). If this amount is 0 or not enough to satisfy the GC, then the process will abort. If this amount is sufficient, user code will continue running, and if it continues allocating it might trigger another OOM without giving back control the event loop.
4 different strategies are implemented to export the heap profile:
- dump it in a folded stack format on stderr: that's the poor man solution that is almost guaranteed to work.
- schedule a JS callback with uv_async: this callback might never be invoked before the process goes OOM again.
- schedule a callback through RequestInterrupt and call JS code from here: this seems to work, but from v8 documentation, an Isolate should not be reentered from RequestInterrupt, hence it might not be a safe thing to do. And since profile export is asynchronous, it might not complete before the process OOM again.
- dump the heap profile as a JSON temp file and spawn a new node process that exports the file: this seems to me the safest method and most likely to work. A new uv_loop is created to enable use of `uv_fs_mkstemp` (for libuv >= 1.34.0) and `uv_spawn`.