Work around faulty file path in webpack sourcemaps
With nextjs/webpack, when there are subdirectories in `pages` directory,
the generated source maps do not reference correctly the generated files
in their `file` property.
For example if the generated file / source maps have paths:
<root>/pages/sub/foo.js(.map)
foo.js.map will have ../pages/sub/foo.js as `file` property instead of
../../pages/sub/foo.js
To workaround this, check first if file referenced in `file` property
exists and if it does not, check if generated file exists alongside the
source map file.